fix: report errno safely from any thread, and clear the clang-tidy sweep
The previous commit left the tidy target red with 99 findings. This clears
all of them, so it passes with WarningsAsErrors '*' across 99 translation
units, and the 697 tests still pass.

One finding was a real defect. std::strerror returns a pointer into a single
process-wide static buffer, and the sites using it report failures from the
32-worker HTTP pool and the procedure threads, not only from single-threaded
startup — so two concurrent failures could overwrite each other's text and
one request could report a reason belonging to another. Since these messages
are what a bus or startup failure is diagnosed from, a misattributed one
sends the diagnosis the wrong way. Adds mm::core::errnoMessage() over
std::system_category().message() — an owned string implemented on the
reentrant strerror_r/strerror_s — and moves platform.cc and
soem_fieldbus_driver.cc onto it. apps/playground keeps std::strerror behind
a NOLINT with the reason at the call site: it is a single-threaded scratch
binary, and reaching for the helper would make it link mm::core for one
message.

resolveSiiCategoryType now matches a raw EEPROM word against a table of
enumerator values instead of casting first and asking afterwards, so no
SiiCategoryType is ever formed from a word that is not an enumerator. The
comment records what this is and is not, because the flattering reading is
wrong: SiiCategoryType fixes its underlying type (: uint16_t), so the
previous cast was well defined and value-preserving and the old switch
returned identical answers. This is hygiene at a validator boundary, not a
bug fix, and it costs something — a switch over the enum is exhaustiveness
checked by -Wswitch and the table is not, so a new single-value category
must be added to the table by hand.

clang-analyzer-optin.core.EnumCastOutOfRange, which flagged the above, is
now excluded. Its only remaining reports are inside uWebSockets' TopicTree.h
and clang-tidy's header filters do not apply to static-analyzer diagnostics
— neither ExcludeHeaderFilterRegex nor --exclude-header-filter suppresses
them, and the vcpkg include is already -isystem. Two unfixable reports would
leave the target permanently red, which is worth more than a check with no
defect to its name here. UBSan's -fsanitize=enum does not stand in for it,
because a fixed-underlying-type enum gives it no undefined behaviour to
trap; the sanitizer plan's note to the contrary is corrected.

The remainder is mechanical: reserve before push loops, const& on
parameters taken by value, const removed where it silently blocked a move,
and redundant optional unwrap-then-rewrap (safe only because the enclosing
if or continue already proves the optional engaged, checked per site).

Twenty NOLINTNEXTLINE markers, each with its reason at the site — mostly
concurrency-mt-unsafe on getenv and std::exit in paths that run before any
thread exists, plus bugprone-exception-escape where the only possible throw
is a standard-library bad_alloc and terminating is the honest outcome, and
performance-trivially-destructible on ~CyclicTimer, which is trivial only in
the build being analysed since the Windows implementation closes a handle
there. Checks that fire a handful of times are left on where the current hit
is defensible; answering a finding at the call site is not a reason to blind
the check.

Two rules about those markers are now in CLAUDE.md, both learned the hard
way. NOLINTNEXTLINE must be the last comment line before the code, since
prose written after it pushes the code out of range and the suppression
silently does nothing while still looking applied. And cpplint parses NOLINT
markers too and rejects a NOLINTEND category it does not know, so a
clang-tidy check name can never appear in the block form.

Adds bugprone-unused-return-value.AllowCastToVoid so (void)f() reads as
deliberate rather than forcing an invented unused variable. Confirmed live
with clang-tidy --dump-config, because CheckOptions has a list form and a
map form depending on version and the unrecognised one is silently ignored
— which a clean run cannot distinguish from an option that never applied.

Also repairs a 207-column line in cmake/lint.cmake that failed cmake-lint,
introduced by the tidy target's own commit, using string(CONCAT) rather
than a disable comment.

+++ libs/core/base64.cc
@@ -42,7 +42,8 @@ std::string base64Encode(std::span<const uint8_t> bytes) {
 
   int written = 0;
   auto* out = reinterpret_cast<unsigned char*>(buffer.data());
-  if (EVP_EncodeUpdate(ctx.get(), out, &written, bytes.data(), bytes.size()) != 1) {
+  if (EVP_EncodeUpdate(ctx.get(), out, &written, bytes.data(), static_cast<int>(bytes.size())) !=
+      1) {
     return {};
   }
   int finalWritten = 0;
